+++ lib/retrospec/exceptions.rb
@@ -1,4 +1,0 @@
-class ParentFoundException < StandardError
-end
-class TypeNotFoundException < StandardError
-end